+++ CHANGELOG.md
@@ -4,6 +4,15 @@ Copyright (c) 2022, GM Consult Pty Ltd
 All rights reserved. 
 -->
 
+### 0.3.1
+
+- **PRE-RELEASE**
+- Added `JSON` and `JsonCollection` type aliases.
+- Added and implemented `TextIndexer.indexCollection` method.
+- Updated tests.
+- Updated examples.
+- Updated documentation.
+
 ### 0.3.0
 
 - **PRE-RELEASE, BREAKING CHANGES**